requirements.txt
Django==1.9.2
argparse==1.2.1
django-tastypie==0.13.1
djangorestframework==3.3.2
python-dateutil==2.4.2
python-mimeparse==1.5.1
six==1.10.0
wsgiref==0.1.2
